refactor(FactoryCreator):删除FactoryCreator API,去除控制台警告信息
