+++ src/components/instant-search/index.js
@@ -5,12 +5,23 @@
 import angular from 'angular';
 import ngResource from 'angular-resource';
 
-import {FactoryCreator} from 'angular-es-utils';
+import template from './instant-search.tpl.html';
+import InstantSearchCtrl from './InstantSearchCtrl';
+import instantSearchLink from './instantSearchLink';
 
-import InstantSearch from './InstantSearch';
+const ddo = {
+	template,
+	controller: InstantSearchCtrl,
+	controllerAs: '$ctrl',
+	link: instantSearchLink,
+	scope: {
+		selectedItem: '=model',
+		options: '<'
+	}
+};
 
 export default angular
 	.module('ccms.components.instantSearch', [ngResource])
-	.directive('instantSearch', FactoryCreator.create(InstantSearch))
+	.directive('instantSearch', () => ddo)
 	.name;
